# openapi.json
{"openapi":"3.1.0","info":{"title":"Todo API","description":"API for managing todo tasks with user isolation","version":"1.0.0"},"paths":{"/auth/register":{"post":{"tags":["authentication"],"summary":"Register User","description":"Register a new user and return a JWT token.","operationId":"register_user_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithTokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["authentication"],"summary":"Login User","description":"Authenticate user and return a JWT token.","operationId":"login_user_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithTokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/token/refresh":{"post":{"tags":["authentication"],"summary":"Refresh Token","description":"Refresh JWT token (placeholder - in a real app, you'd use refresh tokens).","operationId":"refresh_token_auth_token_refresh_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/me":{"get":{"tags":["authentication"],"summary":"Get Current User","description":"Get current authenticated user's information.","operationId":"get_current_user_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/{user_id}/":{"get":{"tags":["chat"],"summary":"Get User Conversations","description":"Get all conversations for the authenticated user.","operationId":"get_user_conversations_api__user_id___get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ConversationRead"},"type":"array","title":"Response Get User Conversations Api  User Id   Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["chat"],"summary":"Create Conversation","description":"Create a new conversation for the authenticated user.","operationId":"create_conversation_api__user_id___post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/{user_id}/{conversation_id}":{"get":{"tags":["chat"],"summary":"Get Conversation","description":"Get a specific conversation by ID for the authenticated user.","operationId":"get_conversation_api__user_id___conversation_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"conversation_id","in":"path","required":true,"schema":{"type":"integer","title":"Conversation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/{user_id}/{conversation_id}/messages":{"post":{"tags":["chat"],"summary":"Add Message To Conversation","description":"Add a message to a specific conversation.","operationId":"add_message_to_conversation_api__user_id___conversation_id__messages_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"conversation_id","in":"path","required":true,"schema":{"type":"integer","title":"Conversation Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessageRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["chat"],"summary":"Get Conversation Messages","description":"Get all messages for a specific conversation.","operationId":"get_conversation_messages_api__user_id___conversation_id__messages_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"conversation_id","in":"path","required":true,"schema":{"type":"integer","title":"Conversation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatMessageRead"},"title":"Response Get Conversation Messages Api  User Id   Conversation Id  Messages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/{user_id}/chat":{"post":{"tags":["chat"],"summary":"Chat Endpoint","description":"Main chat endpoint for AI interaction.\nExpects: {\"message\": \"...\", \"conversation_id\": \"...\"}","operationId":"chat_endpoint_api__user_id__chat_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"Chat Request"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Chat Endpoint Api  User Id  Chat Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["chat"],"summary":"Get User Chats","description":"Get all conversations for the authenticated user.","operationId":"get_user_chats_api__user_id__chat_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get User Chats Api  User Id  Chat Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/{user_id}/chat/{conversation_id}":{"get":{"tags":["chat"],"summary":"Get Chat Conversation","description":"Get a specific conversation by ID.","operationId":"get_chat_conversation_api__user_id__chat__conversation_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"conversation_id","in":"path","required":true,"schema":{"type":"string","title":"Conversation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/users/{user_id}/tasks":{"get":{"summary":"Get Tasks","description":"Get all tasks for the authenticated user","operationId":"get_tasks_api_users__user_id__tasks_get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Task","description":"Create a new task for the authenticated user","operationId":"create_task_api_users__user_id__tasks_post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreateRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/{user_id}/tasks/{id}":{"get":{"summary":"Get Task","description":"Get a specific task by ID for the authenticated user","operationId":"get_task_api_users__user_id__tasks__id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Task","description":"Update a specific task for the authenticated user","operationId":"update_task_api_users__user_id__tasks__id__put","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Task","description":"Delete a specific task for the authenticated user","operationId":"delete_task_api_users__user_id__tasks__id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/{user_id}/tasks/{id}/complete":{"patch":{"summary":"Update Task Completion","description":"Update the completion status of a task for the authenticated user","operationId":"update_task_completion_api_users__user_id__tasks__id__complete_patch","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCompleteRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatMessageCreate":{"properties":{"role":{"type":"string","maxLength":20,"title":"Role"},"content":{"type":"string","title":"Content"},"conversation_id":{"type":"integer","title":"Conversation Id"},"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["role","content","conversation_id","user_id"],"title":"ChatMessageCreate"},"ChatMessageRead":{"properties":{"role":{"type":"string","maxLength":20,"title":"Role"},"content":{"type":"string","title":"Content"},"conversation_id":{"type":"integer","title":"Conversation Id"},"user_id":{"type":"integer","title":"User Id"},"id":{"type":"integer","title":"Id"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"status":{"type":"string","title":"Status"}},"type":"object","required":["role","content","conversation_id","user_id","id","timestamp","status"],"title":"ChatMessageRead"},"ConversationCreate":{"properties":{"title":{"type":"string","maxLength":255,"title":"Title","default":"New Conversation"},"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["user_id"],"title":"ConversationCreate"},"ConversationRead":{"properties":{"title":{"type":"string","maxLength":255,"title":"Title","default":"New Conversation"},"user_id":{"type":"integer","title":"User Id"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"messages":{"items":{},"type":"array","title":"Messages","default":[]}},"type":"object","required":["user_id","id","created_at","updated_at"],"title":"ConversationRead"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TaskCompleteRequest":{"properties":{"completed":{"type":"boolean","title":"Completed"}},"type":"object","required":["completed"],"title":"TaskCompleteRequest"},"TaskCreateRequest":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["title"],"title":"TaskCreateRequest"},"TaskListResponse":{"properties":{"tasks":{"items":{"$ref":"#/components/schemas/TaskResponse"},"type":"array","title":"Tasks"}},"type":"object","required":["tasks"],"title":"TaskListResponse"},"TaskResponse":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"id":{"type":"integer","title":"Id"},"completed":{"type":"boolean","title":"Completed"},"user_id":{"type":"integer","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["title","id","completed","user_id","created_at","updated_at"],"title":"TaskResponse"},"TaskUpdateRequest":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"completed":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Completed"}},"type":"object","title":"TaskUpdateRequest"},"UserCreateRequest":{"properties":{"email":{"type":"string","title":"Email"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","username","password"],"title":"UserCreateRequest"},"UserLoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLoginRequest"},"UserResponse":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"email":{"type":"string","title":"Email"},"username":{"type":"string","title":"Username"},"is_active":{"type":"boolean","title":"Is Active","default":true},"is_superuser":{"type":"boolean","title":"Is Superuser","default":false}},"type":"object","required":["email","username"],"title":"UserResponse"},"UserWithTokenResponse":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"email":{"type":"string","title":"Email"},"username":{"type":"string","title":"Username"},"is_active":{"type":"boolean","title":"Is Active"},"is_superuser":{"type":"boolean","title":"Is Superuser"},"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["email","username","is_active","is_superuser","access_token","token_type"],"title":"UserWithTokenResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}